// src/features/editor/types.ts
export const filters = [
  'none',
  'polaroid',
  'sepia',
  'kodachrome',
  'contrast',
  'brightness',
  'greyscale',
  'brownie',
  'vintage',
  'technicolor',
  'pixelate',
  'invert',
  'blur',
  'sharpen',
  'emboss',
  'removecolor',
  'blacknwhite',
  'vibrance',
  'blendcolor',
  'huerotate',
  'resize',
  'saturation',
  'gamma',
] as const;
// ...
export type ActiveTool =
  | 'select'
  | 'shapes'
  | 'text'
  | 'images'
  | 'draw'
  | 'fill'
  | 'stroke-color'
  | 'stroke-width'
  | 'font'
  | 'opacity'
  | 'filter'
  | 'settings'
  | 'ai'
  | 'remove-bg'
  | 'templates';
// ...
export interface Editor {
  savePNG: () => void;
  saveJPG: () => void;
  saveJPEG: () => void;
  saveJSON: () => void;
  loadJSON: (json: string) => void;
  autoZoom: () => void;
  canUndo: () => boolean;
  canRedo: () => boolean;
  getWorkspace: () => fabric.Object | undefined;
  zoomIn: () => void;
  zoomOut: () => void;
  changeBackground: (background: string) => void;
  changeSize: (size: { width: number; height: number }) => void;
  enableDrawingMode: () => void;
  disableDrawingMode: () => void;
  onUndo: () => void;
  onRedo: () => void;
  onCopy: () => void;
  onPaste: () => void;
  changeImageFilter: (effect: (typeof filters)[number]) => void;
  addImage: (imageUrl: string) => void;
  delete: () => void;
  changeOpacity: (opacity: number) => void;
  bringForward: () => void;
  sendBackwards: () => void;
  changeFontSize: (fontSize: number) => void;
  changeTextAlign: (textAlign: string) => void;
  changeFontUnderline: (underline: boolean) => void;
  changeFontLinethrough: (linethrough: boolean) => void;
  changeFontStyle: (fontStyle: string) => void;
  changeFontWeight: (fontWeight: number) => void;
  changeFontFamily: (fontFamily: string) => void;
  changeFillColor: (color: string) => void;
  changeStrokeColor: (color: string) => void;
  changeStrokeWidth: (width: number) => void;
  changeStrokeDashArray: (strokeDashArray: number[]) => void;
  addText: (text: string, options?: ITextboxOptions) => void;
  addCircle: () => void;
  addSoftRectangle: () => void;
  addRectangle: () => void;
  addTriangle: () => void;
  addInverseTriangle: () => void;
  addDiamond: () => void;
  addPentagon: () => void;
  addHexagon: () => void;
  addStar: () => void;
  addHeart: () => void;
  addArrow: () => void;
  getActiveFontSize: () => number;
  getActiveTextAlign: () => string;
  getActiveFontUnderline: () => boolean;
  getActiveFontLinethrough: () => boolean;
  getActiveFontStyle: () => string;
  getActiveFontWeight: () => number;
  getActiveFontFamily: () => string;
  getActiveOpacity: () => number;
  getActiveFillColor: () => string;
  getActiveStrokeColor: () => string;
  getActiveStrokeWidth: () => number;
  getActiveStrokeDashArray: () => number[];

  canvas: fabric.Canvas;
  selectedObjects: fabric.Object[];
}